Init Seat component test framework
