+++ app/components/Seat/tests/Seat.test.js
@@ -1,0 +1,32 @@
+/**
+* Created by jzobro 20170524
+*/
+// import React from 'react';
+// import { shallow } from 'enzyme';
+// import Seat from '../Seat';
+
+describe('components.seat.seat', () => {
+  describe('seat status', () => {
+    // given props/state
+    // what to look for el = shallow();
+    // expect();
+  });
+
+  describe('info wrapper', () => {
+    // given props/state
+    // what to look for el = shallow();
+    // expect();
+  });
+
+  describe('last action status', () => {
+    // given props/state
+    // what to look for el = shallow();
+    // expect();
+  });
+
+  describe('timer', () => {
+    // given props/state
+    // what to look for el = shallow();
+    // expect();
+  });
+});
